Board tweaks and software update
The use of D2 to control the heater was a mistake as it does not
provide PWM, switch to using D3.

The chosen MOSFET was not suitable for driving the heater at full
current and has thus been switched for a more powerful one.

Move the README so that it's more general for the whole project rather
than just the board design and use Markdown.

Add license to source code.

Add features to the software:

- Drive output with PWM
- Utilise screen backlight
- Save set temperature across boots
- Display percentage power output when on
- Test for detached thermistor and turn off output as safety measure
- Allow scrolling of setpoint with long press of buttons
- Log to serial at approx 1Hz

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>